+++ css-packs/css-reset/src/base.tsx
@@ -181,13 +181,6 @@ export default evaluateInner`
     font-weight: normal;
   }
 
-  text {
-    fill: ${
-      /* eslint-disable-next-line @atlaskit/design-system/no-unsafe-design-token-usage */
-      token('color.text.highEmphasis')
-    };
-  }
-
   code,
   kbd {
     font-family: ${codeFontFamily};