+++ design-system/storybook-addon/register-tokens.js
@@ -1,0 +1,2 @@
+/* eslint-disable */
+require('./dist/cjs/tokens/register');