+++ design-system/storybook-addon/src/index.tsx
@@ -1,0 +1,1 @@
+export { withDesignTokens } from './tokens';