+++ notifications/notification-log-client/src/types.ts
@@ -3,14 +3,8 @@ import { RequestServiceOptions } from '@atlaskit/util-service-support';
 export interface NotificationCountResponse {
   count: number;
 }
-
-export interface CountUnseenRequestServiceOptions
-  extends RequestServiceOptions {
-  useV3NotificationsApi?: boolean;
-}
-
 export interface NotificationLogProvider {
   countUnseenNotifications(
-    options?: CountUnseenRequestServiceOptions,
+    options?: RequestServiceOptions,
   ): Promise<NotificationCountResponse>;
 }