+++ confluence/legacy-mobile-macros/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/legacy-mobile-macros
 
+## 1.0.2
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 1.0.1
 
 ### Patch Changes