+++ design-system/table-tree/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/table-tree
 
+## 9.0.15
+
+### Patch Changes
+
+- [`c55c736ecea`](https://bitbucket.org/atlassian/atlassian-frontend/commits/c55c736ecea) - Patch VULN AFP-3486 AFP-3487 AFP-3488 AFP-3489
+
 ## 9.0.14
 
 ### Patch Changes