+++ editor/editor-mobile-bridge/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @atlaskit/editor-mobile-bridge
 
+## 31.0.4
+
+### Patch Changes
+
+- [`c55c736ecea`](https://bitbucket.org/atlassian/atlassian-frontend/commits/c55c736ecea) - Patch VULN AFP-3486 AFP-3487 AFP-3488 AFP-3489
+- Updated dependencies
+
 ## 31.0.3
 
 ### Patch Changes